Cambios: Modificación en los archivos de configuración, ahora son tres y están separados (spring, hibernate y datasource). Incluido un nuevo pool de conexiones que mejora al de hibernate c3po Modificación de los DAOs, ahora se incluyen Genericos Modificación de los mapeos de spring para incluir archivos js, css... Incluisón de nuevass librerías en el pom

+++ SpringHibernateMaven/src/main/java/es/enbinario/controller/DataController.java
@@ -36,13 +36,13 @@ public class DataController {
 	}
 	
 	@RequestMapping("delete")
-	public ModelAndView deleteUser(@RequestParam int id) {
+	public ModelAndView deleteUser(@RequestParam long id) {
 		dataService.deleteRow(id);
 		return new ModelAndView("redirect:list");
 	}
 	
 	@RequestMapping("edit")
-	public ModelAndView editUser(@RequestParam int id,@ModelAttribute Employee employee) {
+	public ModelAndView editUser(@RequestParam long id,@ModelAttribute Employee employee) {
 		Employee employeeObject = dataService.getRowById(id);
 		return new ModelAndView("edit","employeeObject",employeeObject);
 	}
